+++ test/test_api.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python
+from __future__ import unicode_literals
 
 import sys
 # Allow direct execution
@@ -20,20 +21,36 @@ class ServerTest(unittest.TestCase):
         resp = self.app.get(*args, **kargs)
         return json.loads(resp.data.decode(resp.charset))
 
-    def get_video_info(self, url):
-        return self.get_json('/api/info?%s' % compat_urllib_parse.urlencode({'url': url}))
+    def get_video_info(self, url, flatten=True):
+        args = {
+            'url': url,
+            'flatten': flatten,
+        }
+        return self.get_json('/api/info?%s' % compat_urllib_parse.urlencode(args))
 
     def test_TED(self):
         """Test video (TED talk)"""
         test_url = "http://www.ted.com/talks/dan_dennett_on_our_consciousness.html"
-        info = self.get_video_info(test_url)
+        info = self.get_video_info(test_url, False)
         self.assertEqual(info["url"], test_url)
+        video_info = info['info']
+        keys = ['url', 'ext', 'title']
+        for k in keys:
+            self.assertIn(k, video_info)
+            self.assertIsNotNone(video_info[k])
 
     def test_Vimeo(self):
         """Test Vimeo support"""
         test_url = 'http://vimeo.com/56015672'
         info = self.get_video_info(test_url)
         self.assertEqual(info["url"], test_url)
+
+    def test_flatten(self):
+        test_url = 'http://vimeo.com/56015672'
+        info = self.get_video_info(test_url)
+        videos = info['videos']
+        video_info = videos[0]
+        self.assertIsInstance(video_info, dict)
 
     def test_errors(self):
         resp = self.app.get('/api/info?%s' % compat_urllib_parse.urlencode({'url': 'http://www.google.com'}))